LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/evaluation-error-2-expected.txt

CONSOLE MESSAGE: ReferenceError: Cannot access uninitialized variable.
CONSOLE MESSAGE: ReferenceError: Cannot access uninitialized variable.
CONSOLE MESSAGE: ReferenceError: Cannot access uninitialized variable.

PASS Test that ill-founded cyclic dependencies cause ReferenceError during evaluation, which leads to error events on window, and that exceptions are remembered.
